feat: MainController added, code reduced
