+++ src/main/resources/application.properties
@@ -1,3 +1,3 @@
-logging.level.ru.yandex.practicum.filmorate.controllers.FilmController=debug
-logging.level.ru.yandex.practicum.filmorate.controllers.UserController=debug
+logging.level.ru.yandex.practicum.filmorate.controllers.FilmValidationController=debug
+logging.level.ru.yandex.practicum.filmorate.controllers.UserValidationController=debug
